refactor: analysis_transform ファイルを transform に昇格
- *_analysis_transform.rs を *_transform.rs にリネーム（34ファイル）
- レガシーTransform削除により analysis実装 がメイン実装に昇格
- vector_3d_analysis_transform_tests.rs → vector_3d_transform_tests.rs
- VS Code ネスティング設定を更新

ファイル構造の簡素化:
- 長いファイル名 (*_analysis_transform) から短縮形 (*_transform) へ
- Analysis Matrix4x4実装が標準のtransform実装として確立
- 一貫した命名規則によるコードベースのクリーンアップ

効果:
- 開発効率向上: より短く直感的なファイル名
- 標準化: レガシー削除により一つの実装パスに統一
- 保守性: 明確で一貫したファイル構造

技術詳細:
- Analysis Matrix4x4ベースの効率的な変換実装を保持
- Foundation パターンとの統合を維持
- 型安全性と性能を保ったリファクタリング
