+++ .vscode/settings.json
@@ -133,7 +133,7 @@
     "point_2d.rs": "point_2d_tests.rs,point_2d_extensions.rs,point_2d_transform_safe_tests.rs",
     "point_3d.rs": "point_3d_extensions.rs,point_3d_foundation.rs,point_3d_transform_safe_tests.rs",
     "vector_2d.rs": "vector_2d_tests.rs,vector_2d_extensions.rs",
-    "vector_3d.rs": "vector_3d_tests.rs,vector_3d_extensions.rs,vector_3d_foundation.rs,vector_3d_analysis_transform_tests.rs",
+    "vector_3d.rs": "vector_3d_tests.rs,vector_3d_extensions.rs,vector_3d_foundation.rs,vector_3d_transform_tests.rs",
     "ray_2d.rs": "ray_2d_tests.rs,ray_2d_extensions.rs",
     "ray_3d.rs": "ray_3d_tests.rs,ray_3d_extensions.rs,ray_3d_foundation.rs",
     "arc_2d.rs": "arc_2d_tests.rs,arc_2d_collision.rs,arc_2d_containment.rs,arc_2d_extensions.rs,arc_2d_intersection.rs,arc_2d_metrics.rs,arc_2d_sampling.rs",
