+++ model/geo_primitives/src/bbox_2d_transform.rs
