+++ model/geo_primitives/src/circle_2d_transform.rs
