+++ model/geo_primitives/src/circle_3d_transform.rs
