+++ model/geo_primitives/src/conical_solid_3d_transform.rs
