+++ model/geo_primitives/src/cylindrical_solid_3d_transform.rs
