+++ model/geo_primitives/src/cylindrical_surface_3d_transform.rs
