+++ model/geo_primitives/src/ellipse_2d_transform.rs
