+++ model/geo_primitives/src/ellipse_3d_transform.rs
