+++ model/geo_primitives/src/ellipse_arc_2d_transform.rs
