+++ model/geo_primitives/src/ellipse_arc_3d_transform.rs
