+++ model/geo_primitives/src/ellipsoidal_surface_3d_transform.rs
