+++ model/geo_primitives/src/line_segment_3d_transform.rs
