--- model/geo_primitives/src/point_3d_analysis_transform.rs
@@ -1,374 +1,0 @@
-//! Point3D Analysis Matrix/Vector統合変換実装
-//!
-//! Analysis Matrix/Vectorを直接使用した効率的な座標変換
-//! geo_nurbsのmatrix_transformパターンを基盤とする統一実装
-
-use crate::{Point3D, Vector3D};
-use analysis::linalg::{matrix::Matrix4x4, vector::Vector3};
-use geo_foundation::{AnalysisTransform3D, Angle, Scalar, TransformError};
-
-/// Point3D用Analysis Matrix/Vector変換モジュール
-pub mod analysis_transform {
-    use super::*;
-
-    /// Analysis Vector3への変換
-    impl<T: Scalar> From<Point3D<T>> for Vector3<T> {
-        fn from(point: Point3D<T>) -> Self {
-            Vector3::new(point.x(), point.y(), point.z())
-        }
-    }
-
-    /// Analysis Vector3からの変換
-    impl<T: Scalar> From<Vector3<T>> for Point3D<T> {
-        fn from(vector: Vector3<T>) -> Self {
-            Point3D::from_vector(Vector3D::new(vector.x(), vector.y(), vector.z()))
-        }
-    }
-
-    /// 単一点の行列変換
-    pub fn transform_point_3d<T: Scalar>(point: &Point3D<T>, matrix: &Matrix4x4<T>) -> Point3D<T> {
-        let vec: Vector3<T> = (*point).into();
-        let transformed = matrix.transform_point_3d(&vec);
-        transformed.into()
-    }
-
-    /// 複数点の一括行列変換
-    pub fn transform_points_3d<T: Scalar>(
-        points: &[Point3D<T>],
-        matrix: &Matrix4x4<T>,
-    ) -> Vec<Point3D<T>> {
-        let vectors: Vec<Vector3<T>> = points.iter().map(|&p| p.into()).collect();
-        let transformed_vectors = matrix.transform_points_3d(&vectors);
-        transformed_vectors.into_iter().map(|v| v.into()).collect()
-    }
-
-    /// 平行移動行列の生成
-    pub fn translation_matrix_3d<T: Scalar>(translation: &Vector3D<T>) -> Matrix4x4<T> {
-        let translation_vec: Vector3<T> =
-            Vector3::new(translation.x(), translation.y(), translation.z());
-        Matrix4x4::translation_3d(&translation_vec)
-    }
-
-    /// 回転行列の生成（中心点指定版）
-    pub fn rotation_matrix_3d<T: Scalar>(
-        center: &Point3D<T>,
-        axis: &Vector3D<T>,
-        angle: Angle<T>,
-    ) -> Result<Matrix4x4<T>, TransformError> {
-        let axis_vec: Vector3<T> = Vector3::new(axis.x(), axis.y(), axis.z());
-
-        // ゼロベクトルチェック
-        if axis_vec.norm_squared().is_zero() {
-            return Err(TransformError::ZeroVector(
-                "Cannot rotate around zero vector".to_string(),
-            ));
-        }
-
-        let normalized_axis = axis_vec.normalize().map_err(TransformError::ZeroVector)?;
-
-        // 中心点での回転行列（平行移動->回転->逆平行移動）
-        let center_vec = Vector3::new(center.x(), center.y(), center.z());
-        let translate_to_origin = Matrix4x4::translation_3d(&(-center_vec));
-        let rotation = Matrix4x4::rotation_axis(&normalized_axis, angle.to_radians());
-        let translate_back = Matrix4x4::translation_3d(&center_vec);
-
-        Ok(translate_back * rotation * translate_to_origin)
-    }
-
-    /// スケール行列の生成（中心点・個別軸指定版）
-    pub fn scale_matrix_3d<T: Scalar>(
-        center: &Point3D<T>,
-        scale_x: T,
-        scale_y: T,
-        scale_z: T,
-    ) -> Result<Matrix4x4<T>, TransformError> {
-        if scale_x.is_zero() || scale_y.is_zero() || scale_z.is_zero() {
-            return Err(TransformError::InvalidScaleFactor(
-                "Scale factors cannot be zero".to_string(),
-            ));
-        }
-
-        // 中心点でのスケール行列
-        let center_vec = Vector3::new(center.x(), center.y(), center.z());
-        let scale_vec = Vector3::new(scale_x, scale_y, scale_z);
-        let translate_to_origin = Matrix4x4::translation_3d(&(-center_vec));
-        let scale = Matrix4x4::scale_3d(&scale_vec);
-        let translate_back = Matrix4x4::translation_3d(&center_vec);
-
-        Ok(translate_back * scale * translate_to_origin)
-    }
-
-    /// 均等スケール行列の生成（中心点指定版）
-    pub fn uniform_scale_matrix_3d<T: Scalar>(
-        center: &Point3D<T>,
-        scale_factor: T,
-    ) -> Result<Matrix4x4<T>, TransformError> {
-        if scale_factor.is_zero() {
-            return Err(TransformError::InvalidScaleFactor(
-                "Scale factor cannot be zero".to_string(),
-            ));
-        }
-
-        // 中心点での均等スケール行列
-        let center_vec = Vector3::new(center.x(), center.y(), center.z());
-        let translate_to_origin = Matrix4x4::translation_3d(&(-center_vec));
-        let scale = Matrix4x4::uniform_scale_3d(scale_factor);
-        let translate_back = Matrix4x4::translation_3d(&center_vec);
-
-        Ok(translate_back * scale * translate_to_origin)
-    }
-
-    /// 複合変換行列の構築
-    pub fn composite_point_transform_3d<T: Scalar>(
-        translation: Option<&Vector3D<T>>,
-        rotation: Option<(&Point3D<T>, &Vector3D<T>, Angle<T>)>,
-        scale: Option<(T, T, T)>,
-    ) -> Result<Matrix4x4<T>, TransformError> {
-        let mut result = Matrix4x4::identity();
-
-        // スケール適用
-        if let Some((sx, sy, sz)) = scale {
-            let origin = Point3D::origin();
-            let scale_matrix = scale_matrix_3d(&origin, sx, sy, sz)?;
-            result = result * scale_matrix;
-        }
-
-        // 回転適用
-        if let Some((center, axis, angle)) = rotation {
-            let rotation_matrix = rotation_matrix_3d(center, axis, angle)?;
-            result = result * rotation_matrix;
-        }
-
-        // 平行移動適用
-        if let Some(translation) = translation {
-            let translation_matrix = translation_matrix_3d(translation);
-            result = result * translation_matrix;
-        }
-
-        Ok(result)
-    }
-
-    /// 複合変換行列の構築（均等スケール版）
-    pub fn composite_point_transform_uniform_3d<T: Scalar>(
-        translation: Option<&Vector3D<T>>,
-        rotation: Option<(&Point3D<T>, &Vector3D<T>, Angle<T>)>,
-        scale: Option<T>,
-    ) -> Result<Matrix4x4<T>, TransformError> {
-        let mut result = Matrix4x4::identity();
-
-        // 均等スケール適用
-        if let Some(scale_factor) = scale {
-            let origin = Point3D::origin();
-            let scale_matrix = uniform_scale_matrix_3d(&origin, scale_factor)?;
-            result = result * scale_matrix;
-        }
-
-        // 回転適用
-        if let Some((center, axis, angle)) = rotation {
-            let rotation_matrix = rotation_matrix_3d(center, axis, angle)?;
-            result = result * rotation_matrix;
-        }
-
-        // 平行移動適用
-        if let Some(translation) = translation {
-            let translation_matrix = translation_matrix_3d(translation);
-            result = result * translation_matrix;
-        }
-
-        Ok(result)
-    }
-}
-
-/// Point3DでのAnalysisTransform3D実装（geo_foundation統一トレイト）
-impl<T: Scalar> AnalysisTransform3D<T> for Point3D<T> {
-    type Matrix4x4 = Matrix4x4<T>;
-    type Angle = Angle<T>;
-    type Output = Self;
-
-    fn transform_point_matrix(&self, matrix: &Matrix4x4<T>) -> Self {
-        analysis_transform::transform_point_3d(self, matrix)
-    }
-
-    fn translate_analysis(&self, translation: &Vector3<T>) -> Result<Self, TransformError> {
-        // Vector3からVector3Dへの変換
-        let vector3d = Vector3D::new(translation.x(), translation.y(), translation.z());
-        let matrix = analysis_transform::translation_matrix_3d(&vector3d);
-        Ok(self.transform_point_matrix(&matrix))
-    }
-
-    fn rotate_analysis(
-        &self,
-        center: &Self,
-        axis: &Vector3<T>,
-        angle: Angle<T>,
-    ) -> Result<Self, TransformError> {
-        // Vector3からVector3Dへの変換
-        let axis_vector3d = Vector3D::new(axis.x(), axis.y(), axis.z());
-        let matrix = analysis_transform::rotation_matrix_3d(center, &axis_vector3d, angle)?;
-        Ok(self.transform_point_matrix(&matrix))
-    }
-
-    fn scale_analysis(
-        &self,
-        center: &Self,
-        scale_x: T,
-        scale_y: T,
-        scale_z: T,
-    ) -> Result<Self, TransformError> {
-        let matrix = analysis_transform::scale_matrix_3d(center, scale_x, scale_y, scale_z)?;
-        Ok(self.transform_point_matrix(&matrix))
-    }
-
-    fn uniform_scale_analysis(
-        &self,
-        center: &Self,
-        scale_factor: T,
-    ) -> Result<Self, TransformError> {
-        let matrix = analysis_transform::uniform_scale_matrix_3d(center, scale_factor)?;
-        Ok(self.transform_point_matrix(&matrix))
-    }
-
-    fn apply_composite_transform(
-        &self,
-        translation: Option<&Vector3<T>>,
-        rotation: Option<(&Self, &Vector3<T>, Angle<T>)>,
-        scale: Option<(T, T, T)>,
-    ) -> Result<Self, TransformError> {
-        // Vector3をVector3Dに変換（所有権の問題を回避）
-        let translation_vector3d = translation.map(|t| Vector3D::new(t.x(), t.y(), t.z()));
-        let rotation_adapted = rotation.map(|(center, axis, angle)| {
-            (center, Vector3D::new(axis.x(), axis.y(), axis.z()), angle)
-        });
-        let rotation_ref = rotation_adapted.as_ref().map(|(c, v, a)| (*c, v, *a));
-
-        let matrix = analysis_transform::composite_point_transform_3d(
-            translation_vector3d.as_ref(),
-            rotation_ref,
-            scale,
-        )?;
-        Ok(self.transform_point_matrix(&matrix))
-    }
-
-    fn apply_composite_transform_uniform(
-        &self,
-        translation: Option<&Vector3<T>>,
-        rotation: Option<(&Self, &Vector3<T>, Angle<T>)>,
-        scale: Option<T>,
-    ) -> Result<Self, TransformError> {
-        // Vector3をVector3Dに変換（所有権の問題を回避）
-        let translation_vector3d = translation.map(|t| Vector3D::new(t.x(), t.y(), t.z()));
-        let rotation_adapted = rotation.map(|(center, axis, angle)| {
-            (center, Vector3D::new(axis.x(), axis.y(), axis.z()), angle)
-        });
-        let rotation_ref = rotation_adapted.as_ref().map(|(c, v, a)| (*c, v, *a));
-
-        let matrix = analysis_transform::composite_point_transform_uniform_3d(
-            translation_vector3d.as_ref(),
-            rotation_ref,
-            scale,
-        )?;
-        Ok(self.transform_point_matrix(&matrix))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_analysis_translation() {
-        let point = Point3D::new(1.0, 2.0, 3.0);
-        let translation_vector3d = Vector3D::new(1.0, 1.0, 1.0);
-        let translation: Vector3<f64> = translation_vector3d.into();
-
-        let result = point.translate_analysis(&translation).unwrap();
-        assert_eq!(result, Point3D::new(2.0, 3.0, 4.0));
-    }
-
-    #[test]
-    fn test_analysis_rotation() {
-        let point = Point3D::new(1.0, 0.0, 0.0);
-        let center = Point3D::origin();
-        let axis_vector3d = Vector3D::new(0.0, 0.0, 1.0); // Z軸
-        let axis: Vector3<f64> = axis_vector3d.into();
-        let angle = Angle::from_radians(std::f64::consts::PI / 2.0); // 90度
-
-        let result = point.rotate_analysis(&center, &axis, angle).unwrap();
-
-        // 90度Z軸回転で (1,0,0) → (0,1,0)
-        assert!(result.x().abs() < f64::EPSILON);
-        assert!((result.y() - 1.0).abs() < f64::EPSILON);
-        assert!((result.z() - 0.0).abs() < f64::EPSILON);
-    }
-
-    #[test]
-    fn test_analysis_scale() {
-        let point = Point3D::new(1.0, 2.0, 3.0);
-        let center = Point3D::origin();
-        let scale_factor = 2.0;
-
-        let result = point.uniform_scale_analysis(&center, scale_factor).unwrap();
-        assert_eq!(result, Point3D::new(2.0, 4.0, 6.0));
-    }
-
-    #[test]
-    fn test_composite_transform() {
-        let point = Point3D::new(1.0, 0.0, 0.0);
-        let translation_vector3d = Vector3D::new(1.0, 1.0, 1.0);
-        let translation: Vector3<f64> = translation_vector3d.into();
-        let center = Point3D::origin();
-        let axis_vector3d = Vector3D::new(0.0, 0.0, 1.0);
-        let axis: Vector3<f64> = axis_vector3d.into();
-        let angle = Angle::from_radians(std::f64::consts::PI / 2.0);
-        let scale = (2.0, 2.0, 2.0); // 個別スケール値
-
-        let result = point
-            .apply_composite_transform(
-                Some(&translation),
-                Some((&center, &axis, angle)),
-                Some(scale),
-            )
-            .unwrap();
-
-        println!("Result: ({}, {}, {})", result.x(), result.y(), result.z());
-        // 変換順序の実際の確認: 実際の結果 (-2, 4, 2)
-        // Matrix乗算の順序によりスケール→回転→平行移動の順序で適用される
-        const EPSILON: f64 = 1e-10;
-        assert!((result.x() - (-2.0)).abs() < EPSILON);
-        assert!((result.y() - 4.0).abs() < EPSILON);
-        assert!((result.z() - 2.0).abs() < EPSILON);
-    }
-
-    #[test]
-    fn test_multiple_points_transform() {
-        let points = vec![
-            Point3D::new(1.0, 0.0, 0.0),
-            Point3D::new(0.0, 1.0, 0.0),
-            Point3D::new(0.0, 0.0, 1.0),
-        ];
-
-        let translation = Vector3D::new(1.0, 1.0, 1.0);
-        let matrix = analysis_transform::translation_matrix_3d(&translation);
-
-        let results = analysis_transform::transform_points_3d(&points, &matrix);
-
-        assert_eq!(results[0], Point3D::new(2.0, 1.0, 1.0));
-        assert_eq!(results[1], Point3D::new(1.0, 2.0, 1.0));
-        assert_eq!(results[2], Point3D::new(1.0, 1.0, 2.0));
-    }
-
-    #[test]
-    fn test_error_handling() {
-        let point = Point3D::new(1.0, 2.0, 3.0);
-        let center = Point3D::origin();
-
-        // ゼロベクトル軸での回転
-        let zero_axis_vector3d = Vector3D::new(0.0, 0.0, 0.0);
-        let zero_axis: Vector3<f64> = zero_axis_vector3d.into();
-        let angle = Angle::from_radians(1.0);
-        assert!(point.rotate_analysis(&center, &zero_axis, angle).is_err());
-
-        // ゼロスケール
-        assert!(point.uniform_scale_analysis(&center, 0.0).is_err());
-    }
-}
