+++ model/geo_primitives/src/spherical_solid_3d_transform.rs
