+++ model/geo_primitives/src/spherical_surface_3d_transform.rs
