+++ model/geo_primitives/src/torus_solid_3d_transform.rs
