+++ model/geo_primitives/src/torus_surface_3d_transform.rs
