+++ model/geo_primitives/src/triangle_2d_transform.rs
