+++ model/geo_primitives/src/triangle_3d_transform.rs
